requirements.txt
streamlit==1.50.0
google-genai
duckdb
python-dotenv
streamlit-back-camera-input
pandas
Pillow